# feat: teach ALPArray to store validity only in the encoded array & other minor changes