… Cargo.lock
… docs/rust/quickstart.rst
… encodings/alp/Cargo.toml