@danking 2025-01-22
previously: https://github.com/spiraldb/vortex/pull/1951