@danking 2025-01-30
Clickbench Q39 benchmarked as `2.20987` slower, but locally I'm seeing something like 950ms for this PR versus 1000ms for develop. I've rebased and retriggered the benchmarks.

It makes sense the compression would slow down but decompress-only workloads like ClickBench should only get faster because they load less data (no duplicate validity information) and compute on less data.