@danking 2025-01-30
The average of the ratios supports the argument that this PR is slightly better (maybe ~95% of develop's runtime).

| benchmark | arrow average ratio | parquet average ratio | vortex average ratio |
| --- | --- | --- | --- |
| clickbench | - | 0.9998 | 0.9703 |
| TPC-H | 0.9735 | 0.9838 | 0.9330 |