@danking 2025-02-18
Moved here: https://github.com/spiraldb/vortex/pull/2216